+++ aten/src/ATen/core/grad_mode.h
@@ -1,30 +1,10 @@
 #pragma once
 
 #include <c10/macros/Macros.h>
+#include <c10/core/GradMode.h>
 
 namespace at {
-
-struct TORCH_API GradMode {
-  static bool is_enabled();
-  static void set_enabled(bool enabled);
-};
-
-// A RAII, thread local (!) guard that enables or disables grad mode upon
-// construction, and sets it back to the original value upon destruction.
-struct TORCH_API AutoGradMode {
-  AutoGradMode(bool enabled) : prev_mode(GradMode::is_enabled()) {
-    GradMode::set_enabled(enabled);
-  }
-  ~AutoGradMode() {
-    GradMode::set_enabled(prev_mode);
-  }
-  bool prev_mode;
-};
-
-// A RAII, thread local (!) guard that stops future operations from building
-// gradients.
-struct TORCH_API NoGradGuard : public AutoGradMode {
-  NoGradGuard() : AutoGradMode(/*enabled=*/false) {}
-};
-
+  using GradMode = c10::GradMode;
+  using AutoGradMode = c10::AutoGradMode;
+  using NoGradGuard = c10::NoGradGuard;
 }
